Add release signing config and fix Kotlin warnings
- Configure release signing via ASR_* env vars (lazy, only when set)
- Add *.keystore to .gitignore
- Extract Json instance to companion singleton in ExportImpl
- Replace deprecated monthNumber/dayOfMonth with month/day
- Remove redundant .toLong() (toEpochDays returns Long since 0.6+)
- Add installRelease command to AGENTS.md

+++ androidApp/src/main/java/com/asr/data/database/Converters.kt
@@ -28,6 +28,6 @@ object Converters {
 
     @ColumnTypeConverter
     fun dateToTimestamp(date: LocalDate): Long {
-        return date.toEpochDays().toLong()
+        return date.toEpochDays()
     }
 }
